requirements.txt
PyYAML
pyyaml
requests
atlassian-python-api